+++ builds/liveConfig.js
@@ -13,7 +13,16 @@ const exportObj = {
       importWorkboxFrom: 'local',
       skipWaiting: true,
       clientsClaim: true,
-      navigateFallback: '/index.html'
+      navigateFallback: '/index.html',
+      navigateFallbackBlacklist: [/^\/pages/]
+    },
+    iconPaths: {
+      faviconSVG: 'icons/favicon.svg',
+      favicon32: 'icons/favicon-32x32.png',
+      favicon16: 'icons/favicon-16x16.png',
+      appleTouchIcon: 'icons/apple-touch-icon-152x152.png',
+      maskIcon: 'icons/safari-pinned-tab.svg',
+      msTileImage: 'icons/msapplication-icon-144x144.png'
     }
   },
   chainWebpack: config => {