+++ package-audit.js
@@ -23,7 +23,10 @@ const AUDIT_EXCEPTIONS = [
   'cacheable-request',
   // look into updating after release
   'dns-packet',
-  'decode-uri-component'
+  'decode-uri-component',
+  '@openzeppelin/contracts',
+  'engine.io',
+  'socket.io-parser'
 ];
 const execute = (command, callback) => {
   exec(