+++ src/modules/confirmation/ModuleConfirmation.vue
@@ -317,7 +317,6 @@ export default {
       toDetails: {},
       signing: false,
       links: {
-        ethvm: '',
         explorer: ''
       },
       error: '',
@@ -682,7 +681,6 @@ export default {
       this.toDetails = {};
       this.signing = false;
       this.links = {
-        ethvm: '',
         explorer: ''
       };
       this.error = '';
@@ -729,7 +727,7 @@ export default {
               _this.trackSwap(
                 'swapTxReceivedReceipt',
                 hash,
-                this.network.type.name
+                this.network.type.chainID
               );
             }
           })
@@ -741,7 +739,7 @@ export default {
                 _this.trackSwap(
                   'swapTxBroadcasted',
                   hash,
-                  this.network.type.name
+                  this.network.type.chainID
                 );
               }
               _this.reset();
@@ -776,12 +774,6 @@ export default {
     showSuccess(param) {
       if (isArray(param)) {
         const lastHash = param[param.length - 1].tx.hash;
-        this.links.ethvm = this.network.type.isEthVMSupported.supported
-          ? this.network.type.isEthVMSupported.blockExplorerTX.replace(
-              '[[txHash]]',
-              lastHash
-            )
-          : '';
         this.links.explorer = this.network.type.blockExplorerTX.replace(
           '[[txHash]]',
           lastHash
@@ -797,12 +789,6 @@ export default {
         return;
       }
 
-      this.links.ethvm = this.network.type.isEthVMSupported.supported
-        ? this.network.type.isEthVMSupported.blockExplorerTX.replace(
-            '[[txHash]]',
-            param
-          )
-        : '';
       this.links.explorer = this.network.type.blockExplorerTX.replace(
         '[[txHash]]',
         param
@@ -812,7 +798,7 @@ export default {
         this.trackSwap(
           'swapTransactionSuccessfullySent',
           param,
-          this.network.type.name
+          this.network.type.chainID
         );
       }
     },
@@ -826,7 +812,11 @@ export default {
         event
           .on('transactionHash', res => {
             if (this.isSwap) {
-              this.trackSwap('swapTxBroadcasted', res, this.network.type.name);
+              this.trackSwap(
+                'swapTxBroadcasted',
+                res,
+                this.network.type.chainID
+              );
             }
             this.showTxOverlay = false;
             this.showSuccess(res);
@@ -837,7 +827,7 @@ export default {
               this.trackSwap(
                 'swapTxReceivedReceipt',
                 hash,
-                this.network.type.name
+                this.network.type.chainID
               );
             }
           })
@@ -912,7 +902,7 @@ export default {
                   this.trackSwap(
                     'swapTxReceivedReceipt',
                     hash,
-                    this.network.type.name
+                    this.network.type.chainID
                   );
                 }
               })
@@ -921,9 +911,7 @@ export default {
                   if (rejectedError(e.message)) {
                     this.trackSwap('swapTxRejected');
                     throw new Error(e.message);
-                  }
-
-                  if (i + 1 === this.unsignedTxArr.length) {
+                  } else {
                     this.emitSwapTxFail(e);
                   }
                 }
@@ -933,7 +921,7 @@ export default {
           }
           this.signedTxArray = signed;
         } catch (err) {
-          if (this.isSwap) {
+          if (this.isSwap && !this.isWeb3Wallet) {
             this.emitSwapTxFail(err);
           }
           this.error = errorHandler(err);
@@ -960,12 +948,8 @@ export default {
       if (this.isWeb3Wallet) this.resolver(batchTxEvents);
     },
     emitSwapTxFail(err) {
-      const receipt =
-        err.hasOwnProperty('receipt') &&
-        err.receipt.hasOwnProperty('transactionHash')
-          ? err.receipt.transactionHash
-          : '0x';
-      this.trackSwap('swapTxFailed', receipt, this.network.type.name);
+      const hash = err?.receipt?.transactionHash;
+      this.trackSwap('swapTxFailedV2', hash, this.network.type.chainID);
     },
     btnAction() {
       if (this.isSwap) {