+++ src/views/TheDefaultView.vue
@@ -6,6 +6,10 @@
     </v-main>
     <the-default-footer />
     <the-enkrypt-popup v-if="!isOfflineApp" :show="enkryptLandingPopup" />
+    <the-survey-popup
+      v-if="!isOfflineApp"
+      :show="surveyPopup && !neverShowSurveyPopup"
+    />
   </div>
 </template>
 
@@ -17,10 +21,15 @@ export default {
   components: {
     TheDefaultHeader: () => import('./components-default/TheDefaultHeader'),
     TheDefaultFooter: () => import('./components-default/TheDefaultFooter'),
-    TheEnkryptPopup: () => import('./components-default/TheEnkryptPopup')
+    TheEnkryptPopup: () => import('./components-default/TheEnkryptPopup'),
+    TheSurveyPopup: () => import('./components-default/TheSurveyPopup')
   },
   computed: {
-    ...mapState('popups', ['enkryptLandingPopup']),
+    ...mapState('popups', [
+      'enkryptLandingPopup',
+      'surveyPopup',
+      'neverShowSurveyPopup'
+    ]),
     ...mapState('wallet', ['isOfflineApp'])
   }
 };